+++ facepy/graph_api.py
@@ -1,15 +1,16 @@
-import requests
-
-from .exceptions import FacepyError
-
 try:
     import simplejson as json
 except ImportError:
-    import json
+    import json  # flake8: noqa
+import requests
+
+from urllib import urlencode
+
+from facepy.exceptions import *
+
 
 class GraphAPI(object):
-
-    def __init__(self, oauth_token=False):
+    def __init__(self, oauth_token=False, url='https://graph.facebook.com'):
         """
         Initialize GraphAPI with an OAuth access token.
 
@@ -17,60 +18,76 @@ class GraphAPI(object):
         """
         self.oauth_token = oauth_token
         self.session = requests.session()
+        self.url = url.strip('/')
 
-    def get(self, path='', page=False, **options):
+    def get(self, path='', page=False, retry=3, **options):
         """
         Get an item from the Graph API.
 
         :param path: A string describing the path to the item.
         :param page: A boolean describing whether to return a generator that
                      iterates over each page of results.
+        :param retry: An integer describing how many times the request may be retried.
         :param options: Graph API parameters such as 'limit', 'offset' or 'since'.
 
         See `Facebook's Graph API documentation <http://developers.facebook.com/docs/reference/api/>`_
         for an exhaustive list of parameters.
         """
-
-        response = self._query('GET', path, options, page)
+        response = self._query(
+            method='GET',
+            path=path,
+            data=options,
+            page=page,
+            retry=retry
+        )
 
         if response is False:
-            raise self.FacebookError('Could not get "%s".' % path)
+            raise FacebookError('Could not get "%s".' % path)
 
         return response
 
-    def post(self, path='', **data):
+    def post(self, path='', retry=0, **data):
         """
         Post an item to the Graph API.
 
         :param path: A string describing the path to the item.
-        :param options: Graph API parameters.
+        :param retry: An integer describing how many times the request may be retried.
+        :param data: Graph API parameters such as 'message' or 'source'.
 
         See `Facebook's Graph API documentation <http://developers.facebook.com/docs/reference/api/>`_
         for an exhaustive list of options.
         """
-
-        response = self._query('POST', path, data)
+        response = self._query(
+            method='POST',
+            path=path,
+            data=data,
+            retry=retry
+        )
 
         if response is False:
-            raise self.FacebookError('Could not post to "%s"' % path)
+            raise FacebookError('Could not post to "%s"' % path)
 
         return response
 
-    def delete(self, path):
+    def delete(self, path, retry=3):
         """
         Delete an item in the Graph API.
 
         :param path: A string describing the path to the item.
+        :param retry: An integer describing how many times the request may be retried.
         """
-
-        response = self._query('DELETE', path)
+        response = self._query(
+            method='DELETE',
+            path=path,
+            retry=retry
+        )
 
         if response is False:
-            raise self.FacebookError('Could not delete "%s"' % path)
+            raise FacebookError('Could not delete "%s"' % path)
 
         return response
 
-    def search(self, term, type, page=False, **options):
+    def search(self, term, type, page=False, retry=3, **options):
         """
         Search for an item in the Graph API.
 
@@ -78,6 +95,7 @@ class GraphAPI(object):
         :param type: A string describing the type of items to search for.
         :param page: A boolean describing whether to return a generator that
                      iterates over each page of results.
+        :param retry: An integer describing how many times the request may be retried.
         :param options: Graph API parameters, such as 'center' and 'distance'.
 
         Supported types are ``post``, ``user``, ``page``, ``event``, ``group``, ``place`` and ``checkin``.
@@ -85,8 +103,8 @@ class GraphAPI(object):
         See `Facebook's Graph API documentation <http://developers.facebook.com/docs/reference/api/>`_
         for an exhaustive list of options.
         """
-
         SUPPORTED_TYPES = ['post', 'user', 'page', 'event', 'group', 'place', 'checkin']
+
         if type not in SUPPORTED_TYPES:
             raise ValueError('Unsupported type "%s". Supported types are %s' % (type, ', '.join(SUPPORTED_TYPES)))
 
@@ -95,7 +113,7 @@ class GraphAPI(object):
             'type': type,
         }, **options)
 
-        response = self._query('GET', 'search', options, page)
+        response = self._query('GET', 'search', options, page, retry)
 
         return response
 
@@ -105,68 +123,86 @@ class GraphAPI(object):
 
         :param requests: A list of dictionaries with keys 'method', 'relative_url' and optionally 'body'.
 
-        Yields a list of responses and/or GraphAPI.FacebookError, GraphAPI.HTTPError instances.
+        Yields a list of responses and/or exceptions.
         """
+
+        for request in requests:
+            if 'body' in request:
+                request['body'] = urlencode(request['body'])
+
         responses = self.post(
-            batch = json.dumps(requests)
+            batch=json.dumps(requests)
         )
 
         for response, request in zip(responses, requests):
-            if response:
-                data = json.loads(response['body'])
-            else:
+
+            # Facilitate for empty Graph API responses.
+            #
+            # https://github.com/jgorset/facepy/pull/30
+            if not response:
                 yield None
                 continue
 
-            if not response['code'] == 200:
-                yield self.FacebookError(
-                    '{type} error on {method} "{path}": {msg}.'.format(
-                        type = data['error']['type'],
-                        method = request['method'],
-                        path = request['relative_url'],
-                        msg = data['error']['message']
-                    ), data['error']['code']
-                )
+            try:
+                yield self._parse(response['body'])
+            except FacepyError as exception:
+                exception.request = request
+                yield exception
 
-                continue
+    def fql(self, query, retry=3):
+        """
+        Use FQL to powerfully extract data from Facebook.
 
-            yield data
+        :param query: A FQL query or FQL multiquery ({'query_name': "query",...})
+        :param retry: An integer describing how many times the request may be retried.
 
-    def _query(self, method, path, data={}, page=False):
+        See `Facebook's FQL documentation <http://developers.facebook.com/docs/reference/fql/>`_
+        for an exhaustive list of details.
+        """
+        return self._query(
+            method='GET',
+            path='fql?%s' % urlencode({'q': query}),
+            retry=retry
+        )
+
+    def _query(self, method, path, data=None, page=False, retry=0):
         """
         Fetch an object from the Graph API and parse the output, returning a tuple where the first item
         is the object yielded by the Graph API and the second is the URL for the next page of results, or
         ``None`` if results have been exhausted.
 
         :param method: A string describing the HTTP method.
-        :param url: A string describing the URL.
+        :param path: A string describing the object in the Graph API.
         :param data: A dictionary of HTTP GET parameters (for GET requests) or POST data (for POST requests).
         :param page: A boolean describing whether to return an iterator that iterates over each page of results.
+        :param retry: An integer describing how many times the request may be retried.
         """
+        data = data or {}
 
         def load(method, url, data):
-            if method in ['GET', 'DELETE']:
+            for key in data:
+                value = data[key]
 
-                try:
+                if isinstance(value, (list, dict, set)):
+                    data[key] = json.dumps(value)
+
+            try:
+                if method in ['GET', 'DELETE']:
                     response = self.session.request(method, url, params=data, allow_redirects=True)
-                except requests.RequestException as exception:
-                    raise self.HTTPError(exception.message)
 
+                if method in ['POST', 'PUT']:
+                    files = {}
 
-            if method in ['POST', 'PUT']:
-                files = {}
+                    for key in data:
+                        if hasattr(data[key], 'read'):
+                            files[key] = data[key]
 
-                for key in data:
-                    if hasattr(data[key], 'read'):
-                        files[key] = data[key]
+                    for key in files:
+                        data.pop(key)
 
-                for key in files:
-                    data.pop(key)
-
-                try:
                     response = self.session.request(method, url, data=data, files=files)
-                except requests.RequestException as exception:
-                    raise self.HTTPError(exception.message)
+            except requests.RequestException as exception:
+                raise HTTPError(exception.message)
 
             result = self._parse(response.content)
 
@@ -183,27 +219,35 @@ class GraphAPI(object):
 
                 # Reset pagination parameters.
                 for key in ['offset', 'until', 'since']:
-                    try:
+                    if key in data:
                         del data[key]
-                    except KeyError:
-                        pass
 
                 yield result
 
         # Convert option lists to comma-separated values.
         for key in data:
-            if isinstance(data[key], list) and all([isinstance(item, basestring) for item in data[key]]):
+            if isinstance(data[key], (list, set, tuple)) and all([isinstance(item, basestring) for item in data[key]]):
                 data[key] = ','.join(data[key])
 
-        url = 'https://graph.facebook.com/%s' % path
+        # Support absolute paths too
+        if not path.startswith('/'):
+            path = '/' + str(path)
+
+        url = '%s%s' % (self.url, path)
 
         if self.oauth_token:
             data['access_token'] = self.oauth_token
 
-        if page:
-            return paginate(method, url, data)
-        else:
-            return load(method, url, data)[0]
+        try:
+            if page:
+                return paginate(method, url, data)
+            else:
+                return load(method, url, data)[0]
+        except FacepyError:
+            if retry:
+                return self._query(method, path, data, page, retry - 1)
+            else:
+                raise
 
     def _parse(self, data):
         """
@@ -211,7 +255,6 @@ class GraphAPI(object):
 
         :param data: A string describing the Graph API's response.
         """
-
         try:
             data = json.loads(data)
         except ValueError:
@@ -233,35 +276,25 @@ class GraphAPI(object):
         if type(data) is dict:
             if 'error' in data:
                 error = data['error']
-                error_type = error.get('type')
 
-                if error_type == "OAuthException":
-                    exception = self.OAuthError
+                if error.get('type') == "OAuthException":
+                    exception = OAuthError
                 else:
-                    exception = self.FacebookError
+                    exception = FacebookError
 
                 raise exception(
                     error.get('message'),
                     error.get('code', None)
                 )
+
+            # Facebook occasionally reports errors in its legacy error format.
             if 'error_msg' in data:
-                raise self.FacebookError(
+                raise FacebookError(
                     data.get('error_msg'),
                     data.get('error_code', None)
                 )
 
         return data
 
-    class FacebookError(FacepyError):
-        """Exception for errors returned by the Graph API."""
-
-        def __init__(self, message, code):
-            super(GraphAPI.FacebookError, self).__init__(message)
-
-            self.code = code
-
-    class OAuthError(FacebookError):
-        """Exception for errors specifically related to OAuth."""
-
-    class HTTPError(FacepyError):
-        """Exception for transport errors."""
+    # Proxy exceptions for ease of use and backwards compatibility.
+    FacebookError, OAuthError, HTTPError = FacebookError, OAuthError, HTTPError