+++ HISTORY.rst
@@ -1,6 +1,62 @@
 History
 -------
 
+0.8.4
++++++
+
+* Fixed a bug that caused a KeyError for signed requests that were missing
+  some keys for its "page" attribute.
+
+0.8.3
++++++
+
+* Fixed a bug that caused batch requests with a body to fail.
+* You may now extend access tokens with ``get_extended_access_token``.
+* Fixed a bug that caused paths that started with a slash to fail.
+* Fixed a bug that caused exception messages to be omitted.
+
+0.8.2
++++++
+
+* Facepy exceptions may now be pickled.
+
+0.8.1
++++++
+
+* Fixed a bug that caused pagination to stop prematurely.
+
+0.8.0
++++++
+
+* You may now query application access tokens with ``get_application_access_token``.
+* ``SignedRequest.parse`` now returns a dictionary describing the payload of the signed request
+  instead of a ``SignedRequest`` instance.
+* ``SignedRequest.__init__`` now accepts arguments ``signed_request`` and ``application_secret_key`` and no longer
+  facilitates for constructing arbitrary signed requests.
+* ``SignedRequest#generate`` no longer requires the provision of ``application_secret_key``.
+* ``SignedRequest#oauth_token`` and ``SignedRequest.OAuthToken`` have been removed (deprecated since v0.6).
+* Fixed a bug that caused some exceptions to be returned rather than raised.
+* ``GraphAPI`` now supports retries for ``get``, ``post``, ``delete``, ``search`` and ``fql``.
+* ``GraphAPI#get`` is now more intelligent about pagination and should no longer query Facebook for another page
+  of results if the current page has less elements than ``limit``.
+
+Note: This release is backwards-incompatible.
+
+0.7.0
++++++
+
+* You may now access the original data of the signed request from ``SignedRequest#raw``.
+* You may now issue FQL queries with ``GraphAPI#fql``.
+* Fixed a bug that caused ``GraphAPI#batch`` to crash upon receiving legacy errors from Facebook.
+* ``FacebookError`` exceptions yielded from ``GraphAPI#batch`` now include the request that
+  produced the error.
+
+0.6.9
++++++
+
+* Facepy will now raise ``OAuthError`` for authorization-related errors.
+* Facepy will now reuse the connection to Facebook.
+
 0.6.8
 +++++